requirements.txt
# install with: pip install -r requirements.txt

# Special discord version
discord.py ~= 1.2.5

# database
sqlalchemy ~= 1.3.12
psycopg2-binary ~= 2.8.4
